+++ .gitignore
@@ -2,7 +2,7 @@ HELP.md
 .gradle
 build/
 !gradle/wrapper/gradle-wrapper.jar
-!**/src/main/**
+!**/src/Main/**
 !**/src/test/**
 
 ### STS ###
