+++ src/test/java/blackjack/domain/deck/DeckTest.java
@@ -1,8 +1,10 @@
-package blackjack.domain.card;
+package blackjack.domain.deck;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.deck.Deck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
