+++ src/test/java/blackjack/domain/hand/HandTest.java
@@ -1,7 +1,10 @@
-package blackjack.domain.card;
+package blackjack.domain.hand;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Rank;
+import blackjack.domain.card.Suit;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
